requirements.txt
Pillow==9.0.0
SQLAlchemy==1.4.31
aiohttp==3.8.1
dateparser==1.1.0
genshin@git+https://github.com/thesadru/genshin.py.git@5f9713470b472cd90d1e91dec97723c2f39e00e0#egg=genshin
lxml==4.7.1
py-cord@git+https://github.com/Pycord-Development/pycord.git@70646f2ce8c95379f7367972895886d10f9e4ebb#egg=py-cord
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
rapidfuzz==1.9.1
tenacity==8.0.1
setuptools==60.5.0
psycopg2==2.9.3
git+https://github.com/Rapptz/discord.py
